post_conversation_analysis/requirements.txt
Django>=5.0
djangorestframework
textblob
nltk
django-crontab
sentence-transformers